+++ v4/feature/transfer/src/main/java/exchange/dydx/trading/feature/transfer/status/DydxTransferInstantStatusView.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -202,6 +203,7 @@ object DydxTransferInstantStatusView : DydxComponent {
 
             Text(
                 text = state.title ?: "",
+                textAlign = TextAlign.Center,
                 style = TextStyle.dydxDefault
                     .themeColor(ThemeColor.SemanticColor.text_primary)
                     .themeFont(fontSize = ThemeFont.FontSize.large, fontType = ThemeFont.FontType.plus),
@@ -212,6 +214,7 @@ object DydxTransferInstantStatusView : DydxComponent {
 
             Text(
                 text = state.subtitle ?: "",
+                textAlign = TextAlign.Center,
                 style = TextStyle.dydxDefault
                     .themeColor(ThemeColor.SemanticColor.text_tertiary),
                 modifier = Modifier
