+++ v4/integration/dydxCartera/build.gradle
@@ -34,7 +34,6 @@ dependencies {
 
     implementation project(path: ':v4:integration:cosmos')
     implementation project(path: ':v4:utilities')
-    implementation 'androidx.navigation:navigation-runtime-ktx:2.8.4'
 
     testImplementation "junit:junit:$junitVersion"
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:$composeVersion"
